Issue #335: move NotebookCache.h to lib/model/notebook folder
